requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
